Run k8s health audit after upgrade/rollback
Make a rpc call to conductor to run the audit for
kubernetes health manually.

This will be used to refresh the alarm that can be raised
during upgrade activation or rollback, when kube-apiserver
is being reconfigured, if the periodic audit is executed
at the same time.

Extra: Fix linter error, newline missing at the end of the
       file patch-scripts/start-scripts/24.09.300/pre-start.sh

Test plan:
PASS: Upgrade activation - stx10->11 - AIO-SX* and AIO-DX
PASS: Upgrade activation rollback - stx11->10 - AIO-SX* and AIO-DX

Story: 2011399
Task: 52618

Depends-on: https://review.opendev.org/c/starlingx/config/+/956367

Change-Id: Ib822831c843984040994ba82e27a3222b23ab33d
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>