Fix Ceph-Mon msgr2 not Enabled During Upgrade
After the changes made to reduce addresses in the mgmt network, the
subcloud upgrade orchestrated through dcmanager fails. This occurs
because the ceph-mon address for msgr2 are not enabled in the monitor
during the mon reconfiguration in the upgrade, generating an 800.001
alarm on the platform and causing a timeout during the orchestration's
alarm check.

To fix this, when the ceph-mon is reconfigured during the upgrade, it
already receives the new address for the v2 protocol (msgr2) in its
monmap, ensuring it is initialized with the new protocol and does not
generate alarms that prevent the subcloud upgrade.

Test plan:

- PASS: SX - Ipv4 and Ipv6 (upgrade/rollback)

Closes-bug: 2119953

Change-Id: I8daf746e90fe04688e355f71beea2162d1366822
Signed-off-by: Gabriel Przybysz Gonçalves Júnior <gabriel.przybyszgoncalvesjunior@windriver.com>