Change update function when deployment is interrupted
Currently the approach used by a function (that fails the
deployment when the software-controller suddenly restart) hangs,
leaving the process in an unexpected state and possibly affecting
other deployment commands.

This commit changes the function to use native DeployState and
DeployHostState classes to update the deployment instead of the
deploy_set_failed script, that is intended to be used only
manually and to recover when the system gets into an unexpected
deployment state.

The function behavior remains the same, only using other
mechanisms to update the deployment status.

Test Plan
PASS: reboot the system or restart the software-controller while
      the deployment is in start/activate/activate-rollback states
      and verify the deployment is set as failed when it comes back
      online
PASS: same as previous test case but on deploy host/host-rollback,
      verify the host deploy state is set to failed, and the
      deployment status is set to failed as well

Closes-bug: 2120338

Change-Id: Ib489ffaadbe4193ace877afd2107bdeb7eef5bc1
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>