Add ceph-mon reconfig to auto-recovery
On stx11 there was a reduction on the range of subnets,
thus requiring, on stx10 to stx11 upgrade, that some
services need to be reconfigured in the process, like
ceph. This, in turn, causes the auto-recovery, added as
part of the LVM snapshot feature for AIO-SX, to fail
to rollback to the from-release, since ceph-mon config
change done in the forward path is not being reverted in
this case.

This commit adds the ceph-mon reconfiguration step
while restoring the LVM snapshots.

This commit also removes the expiration time logic
for the LVM snapshots, since they are not really
needed as the feature is scoped only for upgrades,
for now. The expiration time logic can be readded
in the future, if needed.

Test Plan
PASS: AIO-SX e2e upgrade stx-10 to stx-11
PASS: AIO-SX e2e upgrade stx-10 to stx-11 using VIM
PASS: AIO-SX standard rollback procedure
PASS: AIO-SX standard rollback procedure using VIM
PASS: AIO-SX automatic rollback procedure (boot failure)

Closes-bug: 2121090

Change-Id: I40919740cf84d188897080bbf877cc7392c1d466
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>