Remove controllerfs available state check
If controllerfs logical volume size is resized, the state change
to "Available" happens in sysinv audit, which is run when a
final state of deploy activate is reached.

This commit removes the "available" state check after
lv update in activation script. This will help with
activate KPI as it will reduce retries for state change.

Test Plan:
    Pass: Verify state check is skipped after lv update
          when activate script runs

Closes-Bug: 2120362
Change-Id: I6df53f9e984afa84b1aef506090a1289e69d85cb
Signed-off-by: sshathee <shunmugam.shatheesh@windriver.com>