Wait for helm-controller pod on fluxcd rollback
After kube-apiserver is restarted during a rollback, in some instances
an issue with the portieris webhook can happen when the helm rollback
is performed before the helm-controller pod is restarting (portieris
webhook detects a change to an unhealthy resource and fails).

To account for that, added a wait for the helm-controller pod status
to become 'Ready' before proceeding with the rollback. If the wait
has a timeout, the method will warn (instead of failing), since there
might be causes for the pod to not become Ready that can be fixed
after the rollback.

Test plan:
PASS: Perform activation rollback with portieris applied.

Story: 2011357
Task: 52711

Change-Id: Id50016dfccd75eabcc62f7f54bf274ba6b593af3
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>