Inservice patch [software agent]
The /etc merge of a patch applied on top of an inservice patch is
wrongly done. Instead of using the changes from the inservice patch, it
was using the changes from the last reboot required patch.

This commit fixes it by redoing the etc merge in case it detects a
patch being applied on top of an inservice path.

The etc merge is a 3-way merge it.
- Copy the new patch /usr/etc to the new patch /etc
- Get the diff between the previous patch /usr/etc
  and the previous patch /etc
- Apply the diff into the new patch /etc

Test-Plan:
PASS: Apply an inservice patch
 - Check if the Ostree deployment is correctly created
 - Check if the /etc and /usr were bind to the inservice patch
PASS: Apply a patch on top of an inservice patch
 - Check if the Ostree deployment is correctly created
 - Check the logs if the etc merge was done
PASS: Apply 2 patches as reboot required and 2 patches as inservice,
the resulting /etc folder from both should be the same.
PASS: Remove an inservice patch

Story: 2010676
Task: 52777

Change-Id: I64f533d158fb1c987df60a8cb7530e665134fab7
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>