Fix software controller not restarting by flag
PatchControllerMainThread has a mechanism to detect the restart flag and
restart all software controller daemon threads. But this was not working
since the other threads were not receiving the signal to stop, and the
service was not being restarted, but kept hanging.

This commit fixes that, allowing the service to be restarted by closing
all the threads and leaving pmon restart the service.

Test-Plan:
PASS: Apply and remove a patch
PASS: Restart software-controller service with the restart flag
PASS: Restart software-controller service with systemctl

Closes-Bug: 2121976

Change-Id: I1311b171e54cd4bec0ac508c1e6ec23fc5697fae
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>