Inservice post-install script
Adding a post-install for an inservice patch to restart all services
during host install to make the changes live.
If it is the active controller, the software-controller cannot be
restarted at this time, so leaves a flag for it to be restarted during
activate step.

Test-Plan SX/DX:
PASS: Run an inservice patch with this script
      and check if the changes are applied
PASS: Apply an inservice patch pre-bootstrap

Story: 2010676
Task: 52804

Change-Id: Id5b853cea0f80af6124c45aab231c44c34d65576
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>