Create namespace and check fluxcd helmrelease  in advance
This change runs the command to create the flux-helm namespace and also
checks whether fluxcd helmrelease exists in flux-helm namespace.

The purpose of this is to anticipate the fluxcd installation script's
invocation, as the tasks ["Create namespace for FluxCD", "Check if Flux
Helm release already exists"] were taking more than 30 seconds, which
is a long time for simple commands.

With this modification, a decrease of 1 minute was observed in the
execution of the 20-enable-fluxcd-controllers script.

For the improvements noted here to take effect, the review below must
also be merged as well:
https://review.opendev.org/c/starlingx/ansible-playbooks/+/960827

Test Plan:
PASS: build-pkgs && build-image
PASS: Fresh install AIO-SX
PASS: Platform upgrade stx-10 to master
PASS: The create_fluxcd_namespace function in the script eliminates the
      need to run the "Create namespace for FluxCD" task from the
      upgrade-fluxcd-controllers.yml playbook. The function passes to
      ansible whether the namespace was created or not.
PASS: The helm_release_not_exists function pre-checks whether there is a
      helm release installed with the name fluxcd in the flux-helm
      namespace. If the command successfully performs the check, a
      variable is passed to ansible, eliminating the need for the "Check
      if Flux Helm release already exists" task to be executed.

Script reentry tests
PASS: Activation process completes successfully after running it a
      second time after activate fails before Flux is migrated to Helm.
PASS: Activation process completes successfully after running it a
      second time after activate fails after Flux is migrated to Helm
      but before it is upgraded.
PASS: Activation process completes successfully after running it a
      second time after activate fails after Flux is upgraded.

Closes-bug: 2125121

Change-Id: I687cc123bfe66cc615a66fefded71a817a1fb121
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>