Fix release deletion issue in system controller
This commit is to fix a bug where on the system controller,
a non-existed Unavailale release is to be deleted.

The fix is to ensure the Unavailable reelase is to be deleted
when it is N-1 release on the system controller.

Test Plan:

PASS: build and deploy to system controller and sx
PASS: delete N-1, N and N+1 on system controller on stx9
PASS: delete N-1, N and N+1 on system controller on stx10
PASS: delete N and N+1 on sx on stx9
PASS: delete N and N+1 on sx on stx10

Task: 2011357
Story: 52811
Change-Id: Ief740cff87454b1d0e27d18513392d5d50768faa
Signed-off-by: junfeng-li <junfeng.li@windriver.com>