Disable GPG in inactive load import
The major releases loads are creating a remote without the
no-gpg-verify flag, this causes the pull from this repo to fail.

This commit adds this flag when creating an Ostree remote.
Also, it check and add gpg key after load_import script, to fix the
issue in N-1 loads missing this config.

Test-Plan:
PASS: Import a N+1 major release load and see the remote
      was created with gpg disabled in the config file
PASS: Import an N-1 major release load and see the remote
      was created with gpg disabled in the config file

Closes-Bug: 2124981

Change-Id: I6bb4fc6a535d0c8f954d778ec08f18f70a676254
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>