Initialize apt-ostree repo
The system supports uploading patches from different major releases,
but for that, a related apt-ostree instance must exist.

This commit creates the apt-ostree instance before an upload, if it
does not exist.

Test-Plan:
PASS: Upload an N-1 patch without an N-1
      apt-ostree instance should fail
PASS: Upload and apply an N-1 patch with an N-1
      apt-ostree instance should succeed
PASS: Upload and apply an N patch with this fix
PASS: Upload and apply an N+1 patch with an N+1
      apt-ostree instance should succeed
PASS: Upload and apply a patch pre-bootstrap

Closes-Bug: 2125230

Change-Id: If2a387ad9ad437c05b0397a9bf86604180f15c18
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>