Remove kthread paramter to rollback to 24.09
This change is to ensure kernel thread priority parameter is removed
when host rollback to 24.09. The parameter is not being used on 24.09 so
the parameter is removed unconditionally.

Story: 2011401
Task: 52852

Test Plans:
    passed: On lowlatency SX lab, upgrade to 25.09 observed the kernel
parameter rcutree.kthread_prio=21 appares after host unlocked. Abort the
deploy and rollback host, observe that the parameter is removed after
reboot to 24.09.

Signed-off-by: Bin Qian <bin.qian@windriver.com>
Change-Id: I9972ec174e16738af918f27404721fd4f53fc3d1