Check VIM is up after restart
This commit checks if VIM is up post restart, after it points to the
new database file. This is required because if activate rollback
completes before VIM is stable, it misses the notification from
software. This causes VIM to fallback to periodical checking of
software state, adding more time to the overall KPI.

We mitigate this by checking if VIM service is up for 20 s, so
that notification is triggered after that.

Test Plan:
PASS: On AIO-SX verify execution time is reduced and software waits
      for VIM service to come back up

Closes-Bug: 2125580
Change-Id: If6ce82604e2cea93c3d7b8646f35e4e01dbfff1f
Signed-off-by: sshathee <shunmugam.shatheesh@windriver.com>