Script not logging all the reverted apps list in software.log
Executing the 21-k8s-app-upgrade script failed to correctly log the
update status of all applications when the script failed due to a
timeout. This deviates from the fact that the script only triggers
a bulk application update processor in sysinv, and this operation
is reported in batches.

This change analyzes the status of apps in the database when the
timeout expires, reporting the following:
  - Lists of applications that have been updated/reverted.
  - Lists of applications that have not yet started
    updating/reverting.
  - Lists of applications currently in the process of
    updating/reverting.

Test Plan:
PASS: build-pkgs && build-image
PASS: Platform upgrade stx-10 to master
PASS: Platform rollback from master to stx-10
PASS: Simulation of script 21 failing due to a timeout, and all
      applications are logged with their respective statuses.

Closes-bug: 2121827
Depends-On: https://review.opendev.org/c/starlingx/config/+/962259

Change-Id: I805e0341dae6b50cca0e683d502320161d8368b3
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>