Mount /run to chroot environment
This commit is to mount the host /run  to
/sysroot/upgrade/sysroot/run.

USM auth state needs to check the flag controller_config_complete
in the host /run during the data migration in the
chroot environment.

The fix will allow the USM deploy to access the flag.

Test Plan:

PASS: build and deploy on SX
PASS: upgrade stx10 to stx11 with WRO applied
PASS: upgrade stx10 to stx11 without WRO applied

Task: 52872
Story: 2011357

Change-Id: Ie1c53bb0e952de49ce1b414398a290a4491c1af4
Signed-off-by: junfeng-li <junfeng.li@windriver.com>