Force cleanup of checkout directory on deploy start
This commit checks if checkout directory exists on deploy start,
and if it does, then run the cleanup script to remove the directory
forcefully and proceeds with the upgrade, instead of just returning
an error.

The configparser method to get the sw_version is replaced with a
simpler approach.

Two robustness improvements are also added to the cleanup script:
1. Since it needs to determine the to-release major release version
to execute the cleanup tasks (like stopping the to-release database),
but it will fail if the checkout directory is empty or partially
checked out; so in this case, a fallback mechanism will try to get
the major release version from the temporary script path;
2. Check if the major release version to be cleaned up is not the
same as the one running in the system, in case any leftovers of the
previous upgrade are left in the system, so that the deploy cleanup
script doesn't end breaking the current running system.

Test Plan
PASS: deploy start (regression)
PASS: deploy start with checked out to-release ostree-repo,
      verify the cleanup script is executed and deploy start
      proceeds
PASS: deploy start with empty/partially checked out to-release
      ostree-repo, verify the cleanup script is executed and
      deploy start proceeds
PASS: deploy start with checked out ostree-repo from the same
      major release version as the running release, verify the
      cleanup script doesn't execute

Closes-bug: 2126797

Change-Id: I18294c4009da157140aa10091f9a9b3087c9107b
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>