Fix inservice 3-way merge
Ostree 3-way merge always uses the active deployment as base, but in the
cases we have an inservice patch, the inservice deployment is listed as
pending causing the ostree to loose all inservice changes in a new 3-way
merge.

This commit fixes it by point the active deployment folders, used by
ostree 3-way merge, to the pending folders. This way, ostree do the
correctly merge.

Test-Plan:
PASS: Apply a major release on top of an inservice patch
PASS: Apply and remove an inservice patch
PASS: Apply a inservice patch on top of a previous inservice patch

Closes-Bug: 2126705

Change-Id: I11282068d2148673d5225cf97e6a71099e990ac2
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>