update node independently
In rare cases, software-agent is not able to report itself to the
software-controller-daemon, causing HostIpNotFound exception on the
controller, which prevents the software deploy action to the node.

This commit changes to look up inventory hosts for host validation, and
create missing hosts entity for recovery to continue the operation.

This commit also decouples software-agent to report itself (sending
PatchMessageHelloAgentAck msg to software-controller-daemon) from the
PatchMessageHelloAgent query initiated from software-controller-daemon.

Closes-bug: 2127032

Test plan:
    passed: observed software-agent reports itself to
            software-controller-daemon every 10 seconds as expected.
    passed: on a DX system upgrade, do not observed HostIpNotFound
            exception.

Signed-off-by: Bin Qian <bin.qian@windriver.com>
Change-Id: I1fc7e55274ea6a0bf0310a4de996f28edf92073b