Update kernel.env in post-install script
USM has a function to update the kernel.env file after
a new deployment is created, however in case of an in-service
patch, this function is being called before the directories
are remounted, so if the kernel update script is updated in
the patch, the changes won't have effect.

This commit adds a call to the kernel grub update script to
the patch restart script to cover this scenario. A followup
commit will be created to push the call to the function that
updates the kernel file after the remounts are done.

Test Plan
PASS: apply an in-service patch with changes on the puppet
      update grub script, verify it is executed after the
      deploy host operation and that kernel file is updated
      accordingly

Closes-bug: 2127092

Change-Id: I6b7f10051d7be36bedb34c5bdadca34d29e51f53
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>