Disable ceph rook crash alarm during upgrade
During an upgrade, an unexpected daemon crash may occur, and
as a result, ceph may get the 'HEALTH_WARN' status caused by
'X daemon crashed recently'.

This alarm occurs when a process is unexpectedly interrupted
or a failure occurs. However, after restarting the component,
everything returns to normal. However, the Ceph status remains
HEATLH_WARN because the failure is considered recent. This
causes the platform upgrade to fail at stages where alarms
are not expected.

To prevent this behavior during the upgrade, this recent crash
alarm is disabled when 'software deploy start' is executed.
At the end of the upgrade process, when 'software deploy delete'
is executed, it is re-enabled.

Test Plan:
 - PASS: Upgrade on AIO-SX with rook-ceph backend configured.
 - PASS: Upgrade on AIO-SX with ceph backend configured.
 - PASS: Upgrade on STD with rook-ceph backend configured.
 - PASS: Rollback on AIO-SX with rook-ceph backend configured.

Closes-Bug: 2127778

Change-Id: I9a95dfa24f4ce3ea07ea64f4e48da61027dc30b3
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>