Fix cleaning up inactive load import
Clean up load import was wrongly being called and deleting the
/opt/dc-vault/playbooks/MM.mm/, this folder should only be deleted if no
N-1 (MM.mm) release exists anymore.

This commit fixes it by checking if it is the latest N-1 release before
calling this method.

Test-Plan:
PASS: Delete the latest N-1 patch, and check that
      clean up inactive load import was called
PASS: Delete a non-latest N-1 patch, and check that
      clean up inactive load import was not called

Closes-bug: 2128744

Change-Id: Ia7b69ab9ed83c672d39d9fe2248ee35fc20424f9
Signed-off-by: Lindley Vieira <lindley.vieira@windriver.com>