Fix IPv6/IPv4 check during auto-recovery
The IPv6/IPv4 when restoring the ceph mon configuration during the
auto-recovery procedure of the upgrade/rollback process includes a check
that compares None to IPv6Address, which is always False, and the IPv6
is returned as IPv4 (without brackets).

To fix this, the mon_ip (None) was changed to ip_address(content[0]),
which is an IPv4 or IPv6, and the check will work as expected.

Test Plan:
 - PASS: Start an upgrade from an older STX version to a newer one, then
         trigger an auto-recovery during the host unlock step.

NOTE: Test was executed using IPv6 and IPv4

Change-Id: Idf7ebc174f38c4afb5abec1d76f4a8c7f421143f
Signed-off-by: Gustavo Ornaghi Antunes <gustavo.ornaghiantunes@windriver.com>