Remove etcd migration on upgrade deploy start
On stx-11 the etcd database was moved from a versioned
directory /opt/etcd/<major-release> to an unversioned
directory /opt/etcd/db, so the step to create a symlink
to point the to-release to the etcd database during
deploy start is not needed, currently, when upgrading
to stx-12.

This commit removes this step from the script that
prepares the data migration during upgrade deploy start.

Test Plan
PASS: run deploy start successfully, verify no failures
      and verify etcd/k8s is working as expected

Story: 2011357
Task: 52933

Change-Id: Iba4ea9357efbdcbf7f252e94aa036cf2c890a1df
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>