registry.central in dnsmasq local scope
After installing a subcloud using either IPv4 or IPv6, the subcloud must
access registry.central to download container images for applications
such as rook-ceph and others.
To resolve the registry.central hostname, a DNS query is executed.
The corresponding IPv4 or IPv6 address is defined in dnsmasq.addn_conf.

When an application performs a DNS lookup, it typically does not specify
the address family (IPv4 or IPv6).
As a result, the system resolver automatically issues both A (IPv4) and
AAAA (IPv6) queries.
In an IPv4-only installation, the IPv4 query resolves successfully using
the local configuration, but since no IPv6 record exists in
dnsmasq.addn_conf, the IPv6 query is forwarded to the external DNS
(as configured in /etc/resolv.conf).

Ideally, the external DNS should reply with an "NODATA-IPv6" response.
However, if this response is delayed, dropped, or blocked by a network
device, dnsmasq on the subcloud will wait for the reply and retry the
query twice.
This behavior increases the DNS response time for the application and
can lead to installation failures.

Although the issue may appear to originate from the external DNS or
network filtering, the system must not rely on external components for
critical internal name resolutions.
To avoid such problems, DNS queries for registry.central must never be
forwarded to external DNS servers.
For this reason, on subclouds, registry.central is configured within the
local scope of dnsmasq, ensuring it is always resolved locally and never
forwarded externally.

Tests Plan:
for all tests below, execute the following verification:

    - Check if the following line is present in the
     /opt/platform/config/<release>/dnsmasq.addn_conf
     just in the subcloud:
              local=/registry.central/

    - Check in the system service-parameter-list if an entry of
       DNS   local    registry.central
      Is present.

Tests Done:
- DC with AIO-SX subcloud Fresh install
- Upgrade subcloud to new release with this change
- Enrollment install
- subcloud rehoming


Closes-Bug: 2129031

Depends-On: https://review.opendev.org/c/starlingx/config/+/964406
Change-Id: I76be9d793f876ef758548da2551ac973817a3659
Signed-off-by: Fabiano Correa Mercer <fabiano.correamercer@windriver.com>