Progress logs for 21 script improved
This change adjusts script 21's logging during app updates/rollbacks.
The script can now report in real time the list of apps that were
successfully updated/rollbacked during execution.

Previously, the progress log was generated while waiting for the sysinv
API to return the list of apps in batches. This resulted in a
discrepancy between what was updated and what was being reported as
updated.

Now, the query is made via the database, without waiting for the API
to return.

Test Plan:
PASS: build-pkgs && build-image
PASS: Platform upgrade AIO-SX from stx-10 to master
PASS: Platform rollback AIO-SX from master to stx-10
PASS: It is possible to view the progress log reflecting the list of
      applications that completed in real-time update/rollback during
      activation or activation-rollback.

Closes-bug: 2121827

Change-Id: I79932a4db84f0a58f5335a070621d741ae98821b
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>