Create k8s port change flag on rollback
On the stx 11 side, some of sysinv code is executed conditionally
to handle the kube-apiserver port change that was done in the
upgrade from stx 10.

This code is not compatible with a rollback from stx 12, causing
sysinv to use the legacy kube-apiserver port when trying to do
a kube-apiserver health check, resulting in an alarm while the
the rollback isn't done.

To avoid this, this change includes the creation of the mentioned
flag when a rollback is performed, making sysinv use the correct
port.

Test plan:
PASS: Upgrade AIO-SX from stx 11 to stx 12.
PASS: Rollback AIO-SX (before activation) from stx 12 to stx 11.

Closes-bug: 2130905

Change-Id: I4082deea06769bb3215c6071f0ddb263095561ac
Signed-off-by: Marcelo de Castro Loebens <Marcelo.DeCastroLoebens@windriver.com>