Remove the cached deploy state in sync flag
This commit is to remove the cached deploy state in sync flag.

Without using the cached value, the USM will call FM api
to either raise or clear the deploy out of sync alarm
based on the  controller deploy in sync state.

This change is to handle a scenario where the USM
is starting when FM api is not running and the in sync
flag is set to true by mistake. With this change, the in
sync flag will be set correctly after the FM api is up running.

Test Plan:

PASS: build and deploy the iso on DX
PASS: power on/off controller 1 during deploy start to ensure the alarm
          is raised/cleared

Task: 53056
Story: 2011357
Change-Id: I757378239f8aa815f094a46df38b1515a1bdb78e
Signed-off-by: junfeng-li <junfeng.li@windriver.com>