Cleanup obsolete out-of-tree parameters
Existing behaviour - by default out-of-tree-drivers
drivers(ice/i40e/iavf) are loaded and switching to in-tree-drivers
is not supported. So, out_of_tree_drivers service parameter and
multi-drivers-switch/out-of-tree-drivers boot parameters which are
used for switching the drivers are now obsolete.

As part of this commit - ocurences of the above mentioned parameters
are being removed.

TestPlan:
PASS: build-pkgs -c software
PASS: build-image
PASS: bootstrap
PASS: AIO-SX
      - validated no out_of_tree_drivers service parameter exists
      - validated no out-of-tree-drivers kernel parameter exists
      - validated that out_of_tree_drivers service parameter
        cannot be added.

Depends-On: https://review.opendev.org/c/starlingx/config/+/967627

Story: 2011606
Task: 53163

Change-Id: Ifcde3204834feef4e712102f4023efbd38550731
Signed-off-by: Rahul Roshan Kachchap <rahulroshan.kachchap@windriver.com>