Reattempt mechanism for PatchMessageSyncReq
During a software deploy, if a software sync between hosts fails,
PatchMessageSyncReq sends a false positive and does not reattempt
the sync operation, this causes the deploy start operation to fail
silently and wrongly report a start-done status.

This commit adds a reattempt mechanism and only sends a response if
the sync is successful.

Closes-bug: 2133084

Test Plan:
PASS: AIO-SX in-service patch deploy
PASS: AIO-DX in-service patch deploy

Change-Id: If81812173afe0f669ea1a16e26b5b6de5ac198af
Signed-off-by: Matheus Guilhermino <mmachado@windriver.com>