Add OIDC authentication support to USM API
This commit adds OIDC authentication to the USM API, that
now if received a new header 'OIDC-Token' in the request it
will parse the token details and validate it on OIDC dex,
and if the token is valid then the user will be allowed to
run the commands that are enabled for the user roles.

The user roles are mapped as service-parameters in sysinv,
service 'identity', section 'stx' under parameter named
'role-bindings'.

Test Plan
Preparation:
1. Create a LDAP group
2. Create a LDAP user in the created group
3. Set service-parameter role-bindings mapping roles to
   the group (admin,reader,member roles)
4. Run 'service-parameter-apply identity --section stx'
5. Authenticate the user with oidc-auth to create the token

PASS: send a valid oidc token in the request header, verify
      the command executes
PASS: send an invalid oidc token in the request header, verify
      the command is not executed and returns Unauthorized error
PASS: remove the admin role from the group, send a valid oidc
      token in a GET request, verify the command executes
PASS: remove the admin role from the group, send a valid oidc
      token in a POST request, verify the command is not
      executed and returns Unathorized error

Depends-On: https://review.opendev.org/c/starlingx/integ/+/970455

Story: 2011603
Task: 53317

Change-Id: I5e10ce74bea48d55623c364d813571fcd566017f
Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>