Restart systemd-sysctl service in upgrade_hooks
After etc merge sysctl config files are modified and need to be applied
to the running system. The systemd-sysctl service is responsible for
applying the changes to the running kernel parameters.

During upgrade/rollback, if the node is not rebooted any
sysctl config files changes will not be applied until the
systemd-sysctl service is restarted. This change adds
a method to restart the systemd-sysctl service.

Test plan:

PASS - AIO-SX: iso install
               rollback node after host-unlock
               rollback node before host-unlock
               verify the rollback is successful
               complete a full upgrade

Story: 2011583
Task: 53395

Change-Id: I678c0dd755482c8e19e3c32028ece95367da2f69
Signed-off-by: Kyale, Eliud <Eliud.Kyale@windriver.com>